requirements.txt
bcrypt
behave
cryptography
pyjwt
pyopenssl
requests
